requirements.txt

anyio==4.2.0
appdirs==1.4.4
celery==5.3.6
cron-descriptor==1.4.3
Django==5.0.1
django-celery-beat==2.6.0
django-timezone-field==6.1.0
filterpy==1.4.5
paho-mqtt==1.6.1
python-telegram-bot==21.0.1
redis==5.0.2
requests==2.25.1
schedule==1.1.0
supervisor==4.2.2
numpy==1.26.4
